Move `FeedImageLoader` to it's own file
